Rename module bitbucket_pipelines_ssh_key to bitbucket_pipeline_key_pair
